requirements.txt
asgiref==3.8.1
Django==5.0.6
gunicorn==22.0.0
joblib==1.4.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-surprise==1.1.4
scipy==1.14.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
